requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
optuna>=3.3.0

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0

# Data I/O
pyarrow>=12.0.0
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0